+++ CookBook/Models/PlateType.cs
@@ -4,7 +4,7 @@ namespace CookBook.Models
 {
     public class PlateType
     {
-        [Key] public long PlateId { get; set; }
+        [Key] public int PlateId { get; set; }
         public string? PlateName { get; set; }
 
         //Relationships
